5401: move dateinput
